updated menu bar so that it always visible
Going into the smint jquery plugin, I changed some code so that the
window height would be taken into account. When the height is smaller
than 600px, then the menu bar becomes fixed at the top. In this case,
the Home button is replaced with the kanshu title logo. The regular
Kanshu title logo is then hidden. When you scroll
down, the Home button comes back, and is also there when the window is
over 600px tall.